+++ databallpy/utils/constants.py
@@ -36,6 +36,7 @@ DATABALLPY_POSSESSION_TYPES = [
     "penalty",
     "kick_off",
     "goal_kick",
+    "rebound",
     "unspecified",
 ]
 """The possession types that are supported in databallpy."""
@@ -78,3 +79,13 @@ DATABALLPY_PASS_TYPES = [
 """The pass type strings that are supported in databallpy."""
 
 DATABALLPY_POSITIONS = ["goalkeeper", "defender", "midfielder", "forward"]
+
+OPEN_MATCH_IDS_DFL = {
+    "J03WMX": "1. FC Köln vs. FC Bayern München",
+    "J03WN1": "VfL Bochum 1848 vs. Bayer 04 Leverkusen",
+    "J03WPY": "Fortuna Düsseldorf vs. 1. FC Nürnberg",
+    "J03WOH": "Fortuna Düsseldorf vs. SSV Jahn Regensburg",
+    "J03WQQ": "Fortuna Düsseldorf vs. FC St. Pauli",
+    "J03WOY": "Fortuna Düsseldorf vs. F.C. Hansa Rostock",
+    "J03WR9": "Fortuna Düsseldorf vs. 1. FC Kaiserslautern",
+}